move USB camera classes into module
