+++ rosys/vision/usb_camera/__init__.py
@@ -1,0 +1,7 @@
+from .usb_camera import UsbCamera
+from .usb_camera_provider import UsbCameraProvider
+
+__all__ = [
+    'UsbCamera',
+    'UsbCameraProvider',
+]
